+++ src/org/bds/executioner/ExecutionerCluster.java
@@ -217,7 +217,7 @@ public class ExecutionerCluster extends Executioner {
 	}
 
 	@Override
-	protected Cmd createRunCmd(Task task) {
+	public Cmd createRunCmd(Task task) {
 		task.createProgramFile(); // We must create a program file
 
 		if (debug) log("Running task " + task.getId());
