+++ src/org/bds/executioner/ExecutionerSsh.java
@@ -47,7 +47,7 @@ public class ExecutionerSsh extends Executioner {
 	}
 
 	@Override
-	protected Cmd createRunCmd(Task task) {
+	public Cmd createRunCmd(Task task) {
 		task.createProgramFile(); // We must create a program file
 
 		// Create command line
