+++ src/org/bds/task/ShellVmOpcode.java
@@ -1,0 +1,291 @@
+package org.bds.task;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+
+import org.bds.data.Data;
+import org.bds.data.DataRemote;
+import org.bds.executioner.ExecutionerLocal;
+import org.bds.executioner.Executioners;
+import org.bds.lang.expression.ExpressionTask;
+import org.bds.lang.value.Value;
+import org.bds.lang.value.ValueList;
+import org.bds.osCmd.Cmd;
+import org.bds.run.BdsThread;
+
+/**
+ * Execute a 'shell' VM opcode
+ *
+ * A 'shell' is an intermediate between a 'task' and a 'sys':
+ * 	- It executes creating a shell script (like 'task')
+ *  - It logs stdout/stderr (like 'task')
+ *  - It logs exit code (like 'task')
+ *  - Executes immediately (like 'sys')
+ *  - Executed without an executioner (like 'sys')
+ *  - Executes on the local computer (like 'sys')
+ *  - No resource accounting (like 'sys')
+ *
+ * The opcode requires having the following items in the stack
+ *   1) Command to execute
+ *   2) Input dependencies
+ *   3) Output dependencies
+ *
+ * @author pcingola
+ */
+public class ShellVmOpcode extends SysVmOpcode {
+
+	public static String SHELL_EXEC_COMMAND[] = { "bds", "shell" };
+
+	protected Task task;
+	protected TaskDependency taskDependency;
+
+	public ShellVmOpcode(BdsThread bdsThread) {
+		super(bdsThread);
+	}
+
+	/**
+	 * Create command line arguments for "bds -exec"
+	 */
+	protected String[] createBdsExecCmdArgs(Task task) {
+		// Create command line
+		ArrayList<String> args = new ArrayList<String>();
+		for (String arg : SHELL_EXEC_COMMAND)
+			args.add(arg);
+		long timeout = task.getResources().getTimeout() > 0 ? task.getResources().getTimeout() : 0;
+
+		// Add command line parameters for "bds exec"
+		args.add(timeout + ""); // Enforce timeout
+		args.add(task.getStdoutFile() != null ? task.getStdoutFile() : "-"); // Redirect STDOUT to this file
+		args.add(task.getStderrFile() != null ? task.getStderrFile() : "-"); // Redirect STDERR to this file
+		args.add(task.getExitCodeFile() != null ? task.getExitCodeFile() : "-"); // Redirect exit code
+		args.add(task.getProgramFileName()); // Program to execute
+
+		return args.toArray(Cmd.ARGS_ARRAY_TYPE);
+	}
+
+	/**
+	 * Create commands that will be executed in a shell
+	 */
+	protected String createCommands(String sysCmds) {
+		HashMap<String, String> replace = new HashMap<>();
+		StringBuilder sbDown = new StringBuilder();
+		StringBuilder sbUp = new StringBuilder();
+
+		if (taskDependency != null) {
+			//---
+			// Are there any remote inputs?
+			// We need to create the appropriate 'download' commands
+			//---
+			if (taskDependency.getInputs() != null) {
+				for (String in : taskDependency.getInputs()) {
+					Data dataIn = Data.factory(in);
+					if (dataIn.isRemote()) {
+						String uriStr = dataIn.toString();
+						sbDown.append(ExpressionTask.CMD_DOWNLOAD //
+								+ " \"" + uriStr + "\"" //
+								+ " \"" + dataIn.getLocalPath() + "\"" //
+								+ "\n");
+
+						replace.put(uriStr, dataIn.getLocalPath());
+					}
+				}
+			}
+
+			//---
+			// Are there any remote outputs?
+			// We need to create the appropriate 'upload' commands
+			//---
+			if (taskDependency.getOutputs() != null) {
+				for (String out : taskDependency.getOutputs()) {
+					Data dataOut = Data.factory(out);
+					if (dataOut.isRemote()) {
+						String uriStr = dataOut.toString();
+						sbUp.append(ExpressionTask.CMD_UPLOAD //
+								+ " \"" + dataOut.getLocalPath() + "\"" //
+								+ " \"" + uriStr + "\"" //
+								+ "\n");
+
+						replace.put(uriStr, dataOut.getLocalPath());
+
+						// Note, commands executed locally will output to the local file, so
+						// we must make sure that the path exists (otherwise the command
+						// results in an error.
+						((DataRemote) dataOut).mkdirsLocal();
+					}
+				}
+			}
+		}
+
+		// No Down/Up-load? Just return the SYS commands
+		if (sbDown.length() <= 0 && sbUp.length() <= 0) return sysCmds;
+
+		// Replace all occurrences of remote references
+		sysCmds = replace(replace, sysCmds);
+
+		// Put everything together
+		StringBuilder sbSys = new StringBuilder();
+
+		if (sbDown.length() > 0) {
+			sbSys.append("# Download commands\n");
+			sbSys.append(sbDown);
+		}
+
+		sbSys.append(sysCmds);
+
+		if (sbUp.length() > 0) {
+			sbSys.append("\n# Upload commands\n");
+			sbSys.append(sbUp);
+		}
+
+		return sbSys.toString();
+	}
+
+	/**
+	 * Create a task
+	 */
+	protected Task createTask(String sysCmds) {
+		// Get an ID
+		String taskId = taskId();
+
+		// Get commands representing a shell program
+		sysCmds = createCommands(sysCmds);
+
+		// Create Task
+		String sysFileName = getSysFileName(taskId);
+		Task task = new Task(taskId, getBdsNode(), sysFileName, sysCmds);
+
+		// Configure Task parameters
+		task.setVerbose(bdsThread.getConfig().isVerbose());
+		task.setDebug(bdsThread.getConfig().isDebug());
+
+		// Set task options
+		task.setTaskName(getTaskName());
+		task.setCanFail(bdsThread.getBool(ExpressionTask.TASK_OPTION_CAN_FAIL));
+		task.setAllowEmpty(bdsThread.getBool(ExpressionTask.TASK_OPTION_ALLOW_EMPTY));
+		task.setNode(bdsThread.getString(ExpressionTask.TASK_OPTION_NODE));
+		task.setQueue(bdsThread.getString(ExpressionTask.TASK_OPTION_QUEUE));
+		task.setMaxFailCount((int) bdsThread.getInt(ExpressionTask.TASK_OPTION_RETRY) + 1); // Note: Max fail count is the number of retries plus one (we always run at least once)
+		task.setCurrentDir(bdsThread.getCurrentDir());
+
+		// Set task options: Resources
+		task.getResources().setCpus((int) bdsThread.getInt(ExpressionTask.TASK_OPTION_CPUS));
+		task.getResources().setMem(bdsThread.getInt(ExpressionTask.TASK_OPTION_MEM));
+		task.getResources().setWallTimeout(bdsThread.getInt(ExpressionTask.TASK_OPTION_WALL_TIMEOUT));
+		task.getResources().setTimeout(bdsThread.getInt(ExpressionTask.TASK_OPTION_TIMEOUT));
+		if (taskDependency != null) task.setTaskDependency(taskDependency);
+
+		return task;
+	}
+
+	protected TaskDependency createTaskDependency(ValueList outs, ValueList ins) {
+		TaskDependency td = null;
+
+		if (ins.isEmpty() && outs.isEmpty()) return null;
+		td = new TaskDependency(getBdsNode());
+
+		for (Value in : ins)
+			td.addInput(in.asString());
+
+		for (Value out : outs)
+			td.addOutput(out.asString());
+
+		return td;
+	}
+
+	/**
+	 * Dispatch 'shell' for execution
+	 */
+	protected void dispatchTask(Task task) {
+		ExecutionerLocal executioner = (ExecutionerLocal) Executioners.getInstance().get(Executioners.ExecutionerType.LOCAL);
+		Cmd cmd = executioner.createRunCmd(task);
+		cmd.start();
+		try {
+			cmd.join();
+		} catch (InterruptedException e) {
+			throw new RuntimeException(e);
+		}
+	}
+
+	@Override
+	protected String getTaskName() {
+		return bdsThread.hasVariable(ExpressionTask.TASK_OPTION_TASKNAME) ? bdsThread.getString(ExpressionTask.TASK_OPTION_TASKNAME) : null;
+	}
+
+	/**
+	 * Replace all occurrences in 'replace' map
+	 */
+	protected String replace(HashMap<String, String> replace, String sysCmds) {
+		for (String key : replace.keySet()) {
+			String sysCmdsPrev;
+			do {
+				sysCmdsPrev = sysCmds;
+				sysCmds = replace(key, replace.get(key), sysCmds);
+			} while (!sysCmdsPrev.equals(sysCmds)); // Continue while there are replacements
+		}
+
+		return sysCmds;
+	}
+
+	/**
+	 * Replace a single instance of 'oldStr' by 'newStr'
+	 */
+	protected String replace(String oldStr, String newStr, String str) {
+		int start = str.indexOf(oldStr);
+		if (start < 0) return str; // Nothing found
+
+		// Check that 'oldStr' is a separated / quoted word
+		int end = start + oldStr.length();
+		char prevChar = start > 0 ? str.charAt(start - 1) : '\0';
+		char nextChar = end < str.length() ? str.charAt(end) : '\0';
+
+		// Change if surrounded by spaces or quotes
+		boolean change = false;
+		if (prevChar == '\'' && nextChar == '\'') {
+			// Surrounded by single quote
+			change = true;
+		} else if (prevChar == '"' && nextChar == '"') {
+			// Surrounded by double quote
+			change = true;
+		} else if ((prevChar == ' ' || prevChar == '\t' || prevChar == '\n' || prevChar == '\0') && //
+				(nextChar == ' ' || nextChar == '\t' || nextChar == '\n' || nextChar == '\0')) {
+					// Surrounded by space, tab, newline or end_of_string
+					change = true;
+				}
+
+		// Change oldStr by newStr?
+		if (change) { //
+			return str.substring(0, start) // Keep first
+					+ newStr // Replace oldStr by newStr
+					+ (nextChar == '\0' ? "" : str.substring(start + oldStr.length())) // Last part only if there is something after 'oldStr
+			;
+		}
+
+		return str;
+	}
+
+	/**
+	 * Create and run task
+	 */
+	@Override
+	public String run() {
+		// Create task: Get values from stack
+		Value cmd = bdsThread.pop();
+		ValueList ins = (ValueList) bdsThread.pop();
+		ValueList outs = (ValueList) bdsThread.pop();
+
+		taskDependency = createTaskDependency(outs, ins);
+
+		// Create task
+		task = createTask(cmd.asString());
+
+		// Schedule task for execution
+		dispatchTask(task);
+
+		// Push taskId to stack
+		return task.getId();
+	}
+
+	protected String taskId() {
+		return sysId("task");
+	}
+}
